Benchmark C++: fibonacci

+++ src/cpp/algorithm.cpp
@@ -2,7 +2,7 @@
 #include <algorithm>
 #include <cmath>
 
-uint64_t fib(uint32_t n) {
+int fib(int n) {
   if (n <= 1)
     return n;
   return fib(n - 1) + fib(n - 2);
