+++ src/cpp/algorithm.h
@@ -1,7 +1,7 @@
 #include <cstdint>
 #include <queue>
 
-uint64_t fib(uint32_t n);
+int fib(int n);
 int linearSearch(int array[], int length, int search);
 int binarySearch(int sortedArray[], int length, int search);
 void bubbleSort(int array[], int length);
